waylandws_client: only recreate buffer on resize
Currently in case of window resize we return WSEGL_BAD_DRAWABLE
due to which IMG EGL deletes the drawable, and then recreate
the drawable from the native window.
But in the process of recreation the previous buffers are freed
up. Some of these buffers can be held by compositor.
In the new implementation we instead of deleting the whole
drawable only the buffers are recreated when needed.

Signed-off-by: Sunidhi Dixit <sunidhi.dixit@in.bosch.com>